Moving logic from tab controller to tab service

+++ fingerstyle-tab-manager/src/main/java/com/jhmusic/fingerstyletabmanager/service/TabService.java
@@ -1,14 +1,15 @@
 package com.jhmusic.fingerstyletabmanager.service;
 
+import com.jhmusic.fingerstyletabmanager.entity.Difficulty;
 import com.jhmusic.fingerstyletabmanager.entity.Tab;
 import com.jhmusic.fingerstyletabmanager.entity.Tuning;
-import org.springframework.http.ResponseEntity;
+import com.jhmusic.fingerstyletabmanager.util.DifficultyUtils;
 import org.springframework.stereotype.Service;
 import com.jhmusic.fingerstyletabmanager.repository.TabRepository;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,11 +17,11 @@ import java.util.Optional;
 public class TabService {
 
     private final TabRepository tabRepository;
-
-    public TabService(TabRepository tabRepository) {
+    private final TuningService tuningService;
+    public TabService(TabRepository tabRepository, TuningService tuningService) {
         this.tabRepository = tabRepository;
+        this.tuningService = tuningService;
     }
-
     public List<Tab> findAllTabs() {
         return tabRepository.findAll();
     }
@@ -29,12 +30,44 @@ public class TabService {
         return tabRepository.findById(id);
     }
 
-    public Tab saveTab(Tab tab) {
+    public Tab saveTab(String title, String arranger, String tuning, String difficulty, MultipartFile file) throws IOException {
+        // Transform file to byte array if necessary
+        byte[] pdfContent = file.getBytes();
+
+        // Transform the raw string tuning to a Tuning object
+        Tuning tuningObject = tuningService.findTuningByTuningName(tuning);
+
+        System.out.println("id: ");
+        System.out.println(tuningObject.getId());
+        System.out.println(tuningObject.getCounter());
+
+        // Set upload time
+        LocalDateTime uploadDate = LocalDateTime.now();
+
+        // Transform the raw string difficulty to a Difficulty object
+        Difficulty difficultyObject = DifficultyUtils.getDifficultyFromString(difficulty);
+
+        // increment tuning counter by 1 as we create a new tab object
+        tuningObject.incrementCounter();
+        System.out.println(tuningObject.getCounter());
+
+        // Create tab object and save to database
+        Tab tab = new Tab(null, title, arranger, tuningObject, uploadDate, difficultyObject, pdfContent);
         return tabRepository.save(tab);
     }
 
-    public void deleteTabById(Long id) {
-        tabRepository.deleteById(id);
+    public boolean deleteTabById(Long id) {
+        Optional<Tab> optionalTab = tabRepository.findById(id);
+
+        if (optionalTab.isPresent()) {
+            // get tuning from tab and remove 1 from counter as we delete the tab
+            Tuning derivedTuning = optionalTab.get().getTuning();
+            derivedTuning.decrementCounter();
+            tabRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 
 //    public List<Tab> findAllTabsOrderedByAuthor() {
